rag/requirements.txt
# LangChain core and community packages
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
langchain-classic>=1.0.0  # Contains RetrievalQAWithSourcesChain for LangChain 1.0+
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=0.0.6
langchain-mistralai>=0.0.2

# Vector store
faiss-cpu>=1.7.4

# Embeddings (local, no API key needed)
sentence-transformers>=2.2.2
langchain-huggingface>=0.0.1  # For HuggingFaceEmbeddings (LangChain 1.0+)

# Environment variable management
python-dotenv>=1.0.0

# Token counting
tiktoken>=0.5.0

# Utilities
pydantic>=2.0.0

